+++ lib/presentation/screens/splash_screen.dart
@@ -1,9 +1,12 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
+import 'package:flutter_practice/core/constants/strings.dart';
 import 'package:flutter_practice/core/themes/app_theme.dart';
+import 'package:sizer/sizer.dart';
 
 import '../router/app_router.dart';
+import '../utils/app_texts.dart';
 import '../utils/custom_print.dart';
 
 class SplashScreen extends StatefulWidget {
@@ -29,10 +32,18 @@ class _SplashScreenState extends State<SplashScreen> {
     return SafeArea(
       child: Scaffold(
         backgroundColor: AppTheme.backgroundColor,
-        body: const Center(
-          child: CircularProgressIndicator(
-
-          ),
+        body: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Image.asset('assets/images/logo.png'),
+            AppTexts(
+              textString: Strings.appVersion,
+              textFontSize: 10.sp,
+              textAlign: TextAlign.center,
+              fontWeight: FontWeight.bold,
+            ),
+            const CircularProgressIndicator(),
+          ],
         ),
       ),
     );